requirements.txt
airtest-Custom==1.1.3
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
distlib==0.3.0
facebook-wda==0.2.1
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
-irtest-custom==1.1.3
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mss==4.0.3
numpy==1.18.5
opencv-contrib-python==3.4.2.17
Pillow==7.1.2
pip==20.1.1
pylint==2.5.2
pypiwin32==223
pytz==2020.1
pywin32==227
pywinauto==0.6.3
requests==2.23.0
setuptools==41.2.0
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.21
wrapt==1.12.1
zipp==3.1.0
